# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: PrintFlow 3D API
  description: |
    API REST de PrintFlow — Gestion d'impressions 3D et filaments.

    ## Authentification
    Si l'authentification est activée dans les paramètres, ajoutez le header :
    ```
    X-Auth-Token: votre_mot_de_passe
    ```

    ## Intégration Home Assistant
    Exemple de sensor REST pour le stock de filaments :
    ```yaml
    sensor:
      - platform: rest
        name: PrintFlow stock faible
        resource: http://192.168.1.145/api/stats/alerts
        method: GET
        value_template: "{{ value_json | length }}"
        unit_of_measurement: "bobines"
    ```

  version: "1.7.0"
  contact:
    name: PrintFlow 3D
    url: https://github.com/Fabdub59650/printflow-3d
servers:
  - url: /api
    description: Serveur local (Raspberry Pi)
...
paths:
...
  /prints:
    get:
      tags: [Impressions]
      summary: Lister les impressions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [printing, done, failed, paused, queued, cancelled]
        - name: printer_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Liste des impressions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Print'
              example:
                - id: 1
                  name: "Support caméra"
                  status: "done"
                  filament_used: 45
                  actual_duration: 120
                  printer_name: "Elegoo Centauri"
                  filament_name: "PLA Blanc"
                  color_hex: "#FFFFFF"
    post:
      tags: [Impressions]
      summary: Créer une impression
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrintInput'
            example:
              name: "Pied de lampe"
              printer_id: 1
              filament_id: 2
              status: "printing"
              estimated_duration: 180
      responses:
        "201":
          description: Impression créée
  /prints/{id}:
    get:
      tags: [Impressions]
      summary: Détail d'une impression
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        "200":
          description: Détail de l'impression
    put:
      tags: [Impressions]
      summary: Modifier une impression
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrintInput'
      responses:
        "200":
          description: Impression mise à jour
    delete:
      tags: [Impressions]
      summary: Supprimer une impression
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        "200":
          description: Impression supprimée
  /prints/{id}/status:
    patch:
      tags: [Impressions]
      summary: Changer le statut d'une impression
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            example:
              status: "done"
              filament_used: 48
              actual_duration: 135
      responses:
        "200":
          description: Statut mis à jour

  # ── IMPRIMANTES ───────────────────────────────────────────
  /printers:
    get:
      tags: [Imprimantes]
      summary: Lister les imprimantes
      responses:
        "200":
          description: Liste des imprimantes
          content:
            application/json:
              example:
                - id: 1
                  name: "Elegoo Centauri Carbon 2"
                  model: "Centauri Carbon 2"
                  status: "idle"
                  interface_url: "http://192.168.1.200"
    post:
      tags: [Imprimantes]
      summary: Ajouter une imprimante
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: "Bambu X1C"
              model: "X1 Carbon"
              status: "idle"
      responses:
        "201":
          description: Imprimante créée
  /printers/{id}:
    put:
      tags: [Imprimantes]
      summary: Modifier une imprimante
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            example:
              status: "printing"
      responses:
        "200":
          description: Imprimante mise à jour
  /printers/{id}/status:
    patch:
      tags: [Imprimantes]
      summary: Changer le statut d'une imprimante
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            example:
              status: "printing"
      responses:
        "200":
          description: Statut mis à jour

  # ── MAINTENANCE ───────────────────────────────────────────
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Print:
      type: object
      properties:
        id:               { type: integer }
        name:             { type: string }
        status:           { type: string, enum: [printing, done, failed, paused, queued, cancelled] }
        printer_id:       { type: integer }
        filament_id:      { type: integer }
        filament_used:    { type: number, description: "Grammes consommés" }
        estimated_duration: { type: integer, description: "Minutes" }
        actual_duration:    { type: integer, description: "Minutes" }
        progress:         { type: integer, description: "Pourcentage 0-100" }
        file_name:        { type: string }
        notes:            { type: string }
        created_at:       { type: string, format: date-time }
    PrintInput:
      type: object
      required: [name]
      properties:
        name:             { type: string }
        printer_id:       { type: integer }
        filament_id:      { type: integer }
        status:           { type: string }
        estimated_duration: { type: integer }
        actual_duration:    { type: integer }
        filament_used:    { type: number }
        progress:         { type: integer }
        file_name:        { type: string }
        notes:            { type: string }
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: Requis uniquement si l'authentification est activée dans les paramètres
security:
  - ApiKeyAuth: []
  - {}